Imported cemerick.url fork in to gowerstreet
Cemerick URL doesn't support multiple parameters with the same name in a
url. That is, if a url is `www.domain.fr?a=1&a=2` then the library
assigns the parameter `a` to `2` instead of `[1, 2]`.

An unmerged PR supports this behaviour:
https://github.com/henryw374/uri/commit/c22c1ea336af22d5414bd4eab3f03c0a4700d7ce

I've forked the code behind this PR and am adding it to our own maven
repository in S3.