Added support for repeated url parameter names
The url library we use to parse urls in forecast doesn't support url
parameters with the same name showing up repeatedly
(www.domain.com?a=1&a=2). With this commit, the library now does.

This means that we can parse urls with multiple ter_id's, something I
think will be useful if we want to add support for showing pages with a
custom amount of ter ids on them.

Testing:
- Added new unit tests and made sure they passed when running lein test